feat: handle record not found error in RemovePayableUseCase

+++ backend/src/payable/usecases/remove-payable-usecase.ts
@@ -2,12 +2,17 @@ import { Injectable } from '@nestjs/common';
 import { IRemovePayableUseCase } from './remove-payable.usecase.interface';
 import { IPayableRepository } from '../repositories/payable.repository.interface';
 import { RemovePayableInputDTO } from '../dto/remove-payable.input.dto';
+import { RecordNotFoundError } from 'src/persistence/errors/record-not-found.error';
 
 @Injectable()
 export class RemovePayableUseCase implements IRemovePayableUseCase {
   constructor(private readonly payableRepository: IPayableRepository) {}
 
-  execute(removePayableDTO: RemovePayableInputDTO): Promise<void> {
+  async execute(removePayableDTO: RemovePayableInputDTO): Promise<void> {
+    const payable = await this.payableRepository.findById(removePayableDTO);
+    if (!payable) {
+      throw new RecordNotFoundError('Payable');
+    }
     return this.payableRepository.remove(removePayableDTO);
   }
 }
